// frontend/ticket-booking-web/src/app/api/model/paged-result-of-event-dto.ts
/**
 * Ticket Booking Platform API - v1
 *
 * Contact: m1musbal@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EventDto } from './event-dto';


export interface PagedResultOfEventDto { 
    items: Array<EventDto>;
    page: number;
    pageSize: number;
    totalCount: number;
    totalPages?: number;
}